Add ad-hoc patch command for TOML and JSON (#12)
Add the ability to apply a patch to any TOML or JSON file. This is to
allow modifying price-feeder config which has no set home or well-known
directory structure for us to program implicit commands for.

Update readme to reflect the changes and while at it mention the
pre-built binaries for user convenience.